Changed the Node Tree ID from a long to int. Also added a new operator for getting the node id as a combined long from node tree and local id.
git-svn-id: https://svn.apache.org/repos/asf/incubator/vxquery/trunk/vxquery@1468304 13f79535-47bb-0310-9956-ffa450edef68

+++ vxquery-core/src/main/java/org/apache/vxquery/runtime/functions/step/AbstractPathStepScalarEvaluator.java
@@ -223,8 +223,7 @@ public abstract class AbstractPathStepScalarEvaluator extends AbstractTaggedValu
         }
         mainOut.write(header);
         if (hasNodeIds) {
-            // TODO put in real id.
-            mainOut.writeLong(ntp.getRootNodeId());
+            mainOut.writeInt(ntp.getRootNodeId());
         }
         if (hasDictionary) {
             mainOut.write(ntp.getByteArray(), ntp.getDictionaryOffset(), ntp.getDictionarySize());
